update names and move read sat info to utils
