+++ src/plotting/plot_mysql_sensor.py
@@ -64,7 +64,7 @@ for index, row in unique_sensor.iterrows():
 fig = plt.figure(dpi=300)
 fig.patch.set_facecolor('white')
 ax = fig.add_axes([0, 0.1, 1, height+step])
-plt.title("Inventory of Clean Bucket Atmosphere Sensors")
+plt.title("Inventory of NNJA Atmosphere Sensors")
 plt.xlabel('Observation Date')
 plt.ylabel('Sensor')
